+++ app-modules/alert/src/Models/Concerns/HasAlerts.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Assist\Alert\Models\Concerns;
+
+use Assist\Alert\Models\Alert;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
+
+trait HasAlerts
+{
+    public function alerts(): MorphMany
+    {
+        return $this->morphMany(Alert::class, 'concern');
+    }
+}
