+++ app-modules/alert/src/Models/Contracts/Alertable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Assist\Alert\Models\Contracts;
+
+use Illuminate\Database\Eloquent\Relations\MorphMany;
+
+interface Alertable
+{
+    public function alerts(): MorphMany;
+}
