+++ app-modules/assist-data-model/src/Filament/Resources/StudentResource/Pages/ManageStudentTasks.php
@@ -2,148 +2,10 @@
 
 namespace Assist\AssistDataModel\Filament\Resources\StudentResource\Pages;
 
-use Filament\Forms\Form;
-use Filament\Tables\Table;
-use Assist\Task\Models\Task;
-use Assist\Task\Enums\TaskStatus;
-use Filament\Tables\Filters\Filter;
-use Assist\Prospect\Models\Prospect;
-use Filament\Forms\Components\Select;
-use Filament\Tables\Actions\EditAction;
-use Filament\Tables\Columns\TextColumn;
-use Illuminate\Database\Eloquent\Model;
-use App\Filament\Resources\UserResource;
-use Filament\Forms\Components\TextInput;
-use Filament\Tables\Actions\CreateAction;
-use Filament\Tables\Actions\DetachAction;
-use Filament\Tables\Filters\SelectFilter;
-use Assist\AssistDataModel\Models\Student;
-use Filament\Tables\Actions\BulkActionGroup;
-use Filament\Forms\Components\DateTimePicker;
-use Filament\Tables\Actions\DetachBulkAction;
-use Filament\Resources\Pages\ManageRelatedRecords;
-use Assist\Prospect\Filament\Resources\ProspectResource;
 use Assist\AssistDataModel\Filament\Resources\StudentResource;
-use Assist\Task\Filament\Resources\TaskResource\Components\TaskViewAction;
+use Assist\Task\Filament\Resources\TaskResource\Pages\ManageTasks;
 
-class ManageStudentTasks extends ManageRelatedRecords
+class ManageStudentTasks extends ManageTasks
 {
     protected static string $resource = StudentResource::class;
-
-    protected static string $relationship = 'tasks';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $navigationLabel = 'Tasks';
-
-    // TODO: Automatically set from Filament based on relationship name
-    public static ?string $breadcrumb = 'Tasks';
-
-    protected static ?string $navigationIcon = 'heroicon-o-clipboard-document-check';
-
-    public function form(Form $form): Form
-    {
-        return $form
-            ->schema([
-                TextInput::make('description')
-                    ->label('Description')
-                    ->required()
-                    ->string(),
-                DateTimePicker::make('due')
-                    ->label('Due Date')
-                    ->native(false),
-                Select::make('assigned_to')
-                    ->label('Assigned To')
-                    ->relationship('assignedTo', 'name')
-                    ->nullable()
-                    ->searchable(['name', 'email'])
-                    ->default(auth()->id()),
-            ]);
-    }
-
-    public function table(Table $table): Table
-    {
-        return $table
-            ->recordTitleAttribute('description')
-            ->columns([
-                TextColumn::make('description')
-                    ->searchable()
-                    ->wrap()
-                    ->limit(50),
-                TextColumn::make('status')
-                    ->formatStateUsing(fn (TaskStatus $state): string => str($state->value)->title()->headline())
-                    ->badge()
-                    ->color(fn (Task $record) => $record->status->getTableColor()),
-                TextColumn::make('due')
-                    ->label('Due Date')
-                    ->sortable(),
-                TextColumn::make('assignedTo.name')
-                    ->label('Assigned To')
-                    ->url(fn (Task $record) => $record->assignedTo ? UserResource::getUrl('view', ['record' => $record->assignedTo]) : null),
-                TextColumn::make('concern.full')
-                    ->label('Concern')
-                    ->url(fn (Task $record) => match ($record->concern ? $record->concern::class : null) {
-                        Student::class => StudentResource::getUrl('view', ['record' => $record->concern]),
-                        Prospect::class => ProspectResource::getUrl('view', ['record' => $record->concern]),
-                        default => null,
-                    }),
-            ])
-            ->filters([
-                Filter::make('my_tasks')
-                    ->label('My Tasks')
-                    ->query(
-                        fn ($query) => $query->where('assigned_to', auth()->id())
-                    ),
-                SelectFilter::make('assignedTo')
-                    ->label('Assigned To')
-                    ->relationship('assignedTo', 'name')
-                    ->searchable()
-                    ->multiple(),
-                SelectFilter::make('status')
-                    ->label('Status')
-                    ->options(collect(TaskStatus::cases())->mapWithKeys(fn (TaskStatus $direction) => [$direction->value => \Livewire\str($direction->name)->title()->headline()]))
-                    ->multiple()
-                    ->default(
-                        [
-                            TaskStatus::PENDING->value,
-                            TaskStatus::IN_PROGRESS->value,
-                        ]
-                    ),
-            ])
-            ->headerActions([
-                $this->createAction(),
-            ])
-            ->actions([
-                TaskViewAction::make(),
-                EditAction::make(),
-                DetachAction::make(),
-            ])
-            ->recordUrl(null)
-            ->bulkActions([
-                BulkActionGroup::make([
-                    DetachBulkAction::make(),
-                ]),
-            ])
-            ->emptyStateActions([
-                $this->createAction(),
-            ]);
-    }
-
-    protected function createAction()
-    {
-        return CreateAction::make()
-            ->using(function (array $data, string $model): Model {
-                $data = collect($data);
-
-                /** @var Task $task */
-                $task = new ($model)($data->except('assigned_to')->toArray());
-
-                $task->assigned_to = $data->get('assigned_to');
-
-                $task->concern()->associate($this->getOwnerRecord());
-
-                $task->save();
-
-                return $task;
-            });
-    }
 }
