+++ app-modules/engagement/src/Filament/Resources/EngagementResource/Pages/ManageEngagements.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace Assist\Engagement\Filament\Resources\EngagementResource\Pages;
+
+use Illuminate\Database\Eloquent\Model;
+use Filament\Resources\Pages\ManageRelatedRecords;
+use Filament\Resources\RelationManagers\RelationManager;
+use Assist\Prospect\Filament\Resources\ProspectResource\RelationManagers\EngagementsRelationManager;
+use Assist\Prospect\Filament\Resources\ProspectResource\RelationManagers\EngagementResponsesRelationManager;
+
+class ManageEngagements extends ManageRelatedRecords
+{
+    // TODO: Obsolete when there is no table, remove from Filament
+    protected static string $relationship = 'engagements';
+
+    protected static ?string $navigationLabel = 'Engagements';
+
+    protected static ?string $breadcrumb = 'Engagements';
+
+    protected static ?string $navigationIcon = 'heroicon-o-chat-bubble-bottom-center-text';
+
+    public static function canAccess(?Model $record = null): bool
+    {
+        /** @var RelationManager $relationManager */
+        foreach ([
+            EngagementsRelationManager::class,
+            EngagementResponsesRelationManager::class,
+        ] as $relationManager) {
+            if (! $relationManager::canViewForRecord($record, static::class)) {
+                continue;
+            }
+
+            return true;
+        }
+
+        return false;
+    }
+
+    public function getRelationManagers(): array
+    {
+        return [
+            EngagementsRelationManager::class,
+            EngagementResponsesRelationManager::class,
+        ];
+    }
+}
