+++ app-modules/prospect/src/Filament/Resources/ProspectResource.php
@@ -2,10 +2,19 @@
 
 namespace Assist\Prospect\Filament\Resources;
 
+use Filament\Pages\Page;
 use Filament\Resources\Resource;
-use Filament\Resources\Pages\Page;
 use Assist\Prospect\Models\Prospect;
-use Assist\Prospect\Filament\Resources\ProspectResource\Pages;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\EditProspect;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ViewProspect;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ListProspects;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\CreateProspect;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ManageProspectFiles;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ManageProspectTasks;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ManageProspectAlerts;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ManageProspectEngagement;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ManageProspectInteractions;
+use Assist\Prospect\Filament\Resources\ProspectResource\Pages\ManageProspectSubscriptions;
 
 class ProspectResource extends Resource
 {
@@ -20,30 +29,30 @@ class ProspectResource extends Resource
     public static function getRecordSubNavigation(Page $page): array
     {
         return $page->generateNavigationItems([
-            Pages\ViewProspect::class,
-            Pages\EditProspect::class,
-            Pages\ManageProspectEngagement::class,
-            Pages\ManageProspectFiles::class,
-            Pages\ManageProspectAlerts::class,
-            Pages\ManageProspectTasks::class,
-            Pages\ManageProspectSubscriptions::class,
-            Pages\ManageProspectInteractions::class,
+            ViewProspect::class,
+            EditProspect::class,
+            ManageProspectEngagement::class,
+            ManageProspectFiles::class,
+            ManageProspectAlerts::class,
+            ManageProspectTasks::class,
+            ManageProspectSubscriptions::class,
+            ManageProspectInteractions::class,
         ]);
     }
 
     public static function getPages(): array
     {
         return [
-            'index' => Pages\ListProspects::route('/'),
-            'create' => Pages\CreateProspect::route('/create'),
-            'edit' => Pages\EditProspect::route('/{record}/edit'),
-            'manage-alerts' => Pages\ManageProspectAlerts::route('/{record}/alerts'),
-            'manage-engagement' => Pages\ManageProspectEngagement::route('/{record}/engagement'),
-            'manage-files' => Pages\ManageProspectFiles::route('/{record}/files'),
-            'manage-interactions' => Pages\ManageProspectInteractions::route('/{record}/interactions'),
-            'manage-subscriptions' => Pages\ManageProspectSubscriptions::route('/{record}/subscriptions'),
-            'manage-tasks' => Pages\ManageProspectTasks::route('/{record}/tasks'),
-            'view' => Pages\ViewProspect::route('/{record}'),
+            'index' => ListProspects::route('/'),
+            'create' => CreateProspect::route('/create'),
+            'edit' => EditProspect::route('/{record}/edit'),
+            'manage-alerts' => ManageProspectAlerts::route('/{record}/alerts'),
+            'manage-engagement' => ManageProspectEngagement::route('/{record}/engagement'),
+            'manage-files' => ManageProspectFiles::route('/{record}/files'),
+            'manage-interactions' => ManageProspectInteractions::route('/{record}/interactions'),
+            'manage-subscriptions' => ManageProspectSubscriptions::route('/{record}/subscriptions'),
+            'manage-tasks' => ManageProspectTasks::route('/{record}/tasks'),
+            'view' => ViewProspect::route('/{record}'),
         ];
     }
 }
