+++ app-modules/prospect/src/Filament/Resources/ProspectResource/Pages/ManageProspectAlerts.php
@@ -2,20 +2,10 @@
 
 namespace Assist\Prospect\Filament\Resources\ProspectResource\Pages;
 
-use Filament\Resources\Pages\ManageRelatedRecords;
 use Assist\Prospect\Filament\Resources\ProspectResource;
+use Assist\Alert\Filament\Resources\AlertResource\Pages\ManageAlerts;
 
-class ManageProspectAlerts extends ManageRelatedRecords
+class ManageProspectAlerts extends ManageAlerts
 {
     protected static string $resource = ProspectResource::class;
-
-    protected static string $relationship = 'alerts';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $navigationLabel = 'Alerts';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $breadcrumb = 'Alerts';
-
-    protected static ?string $navigationIcon = 'heroicon-o-bell-alert';
 }
