+++ app-modules/prospect/src/Filament/Resources/ProspectResource/Pages/ManageProspectInteractions.php
@@ -2,59 +2,13 @@
 
 namespace Assist\Prospect\Filament\Resources\ProspectResource\Pages;
 
-use Filament\Forms\Form;
-use Filament\Tables\Table;
-use Filament\Infolists\Infolist;
 use Assist\Prospect\Models\Prospect;
-use Filament\Forms\Components\Hidden;
-use Filament\Forms\Components\Component;
-use Filament\Forms\Components\MorphToSelect;
-use Filament\Resources\Pages\ManageRelatedRecords;
 use Assist\Prospect\Filament\Resources\ProspectResource;
-use Assist\Interaction\Filament\Resources\InteractionResource\Pages\CreateInteraction;
-use Assist\Interaction\Filament\Resources\InteractionResource\RelationManagers\HasManyMorphedInteractionsRelationManager;
+use Assist\Interaction\Filament\Resources\InteractionResource\Pages\ManageInteractions;
 
-class ManageProspectInteractions extends ManageRelatedRecords
+class ManageProspectInteractions extends ManageInteractions
 {
     protected static string $resource = ProspectResource::class;
 
-    protected static string $relationship = 'interactions';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $breadcrumb = 'Interactions';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $navigationLabel = 'Interactions';
-
-    protected static ?string $navigationIcon = 'heroicon-o-arrow-path-rounded-square';
-
-    public function form(Form $form): Form
-    {
-        $createInteractionForm = (resolve(CreateInteraction::class))->form($form);
-
-        $formComponents = collect($createInteractionForm->getComponents())->filter(function (Component $component) {
-            if (! $component instanceof MorphToSelect) {
-                return true;
-            }
-        })->toArray();
-
-        return parent::form($createInteractionForm)
-            ->schema([
-                Hidden::make('interactable_id')
-                    ->default($this->getOwnerRecord()->identifier()),
-                Hidden::make('interactable_type')
-                    ->default(resolve(Prospect::class)->getMorphClass()),
-                ...$formComponents,
-            ]);
-    }
-
-    public function infolist(Infolist $infolist): Infolist
-    {
-        return (resolve(HasManyMorphedInteractionsRelationManager::class))->infolist($infolist);
-    }
-
-    public function table(Table $table): Table
-    {
-        return (resolve(HasManyMorphedInteractionsRelationManager::class))->table($table);
-    }
+    protected static string $interactableType = Prospect::class;
 }
