+++ app-modules/prospect/src/Filament/Resources/ProspectResource/Pages/ManageProspectTasks.php
@@ -2,20 +2,10 @@
 
 namespace Assist\Prospect\Filament\Resources\ProspectResource\Pages;
 
-use Filament\Resources\Pages\ManageRelatedRecords;
 use Assist\Prospect\Filament\Resources\ProspectResource;
+use Assist\Task\Filament\Resources\TaskResource\Pages\ManageTasks;
 
-class ManageProspectTasks extends ManageRelatedRecords
+class ManageProspectTasks extends ManageTasks
 {
     protected static string $resource = ProspectResource::class;
-
-    protected static string $relationship = 'tasks';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $navigationLabel = 'Tasks';
-
-    // TODO: Automatically set from Filament based on relationship name
-    protected static ?string $breadcrumb = 'Tasks';
-
-    protected static ?string $navigationIcon = 'heroicon-o-clipboard-document-check';
 }
